move new tests into test folder
